MAJ de la photo de profil sur les commentaires et sur le sigle de bonjour
