+++ src/store/slices/base/constants.ts
@@ -1,2 +1,3 @@
 export const SLICE_BASE_NAME = 'base'
 export const SLICE_CART_NAME = 'cart'
+export const SLICE_AVATAR_URL_NAME = 'avatarUrl'
